Environments Page - Environments sub-component breakout - broke out remaining sub-components and minor import cleanup to remove import * pattern usage. - no functional changes
