+++ src/analysis/Environments/Environments.test.ts
@@ -14,22 +14,19 @@ import {
   generateTestDiskWithGoogleWorkspace,
   generateTestListGoogleRuntime,
 } from 'src/analysis/_testData/testData';
-import {
-  EnvironmentNavActions,
-  Environments,
-  EnvironmentsProps,
-  LeoResourcePermissionsProvider,
-} from 'src/analysis/Environments/Environments';
 import { appToolLabels } from 'src/analysis/utils/tool-utils';
 import { AzureConfig } from 'src/libs/ajax/leonardo/models/runtime-config-models';
 import { Runtime, runtimeStatuses } from 'src/libs/ajax/leonardo/models/runtime-models';
 import { LeoAppProvider } from 'src/libs/ajax/leonardo/providers/LeoAppProvider';
 import { LeoDiskProvider } from 'src/libs/ajax/leonardo/providers/LeoDiskProvider';
 import { LeoRuntimeProvider } from 'src/libs/ajax/leonardo/providers/LeoRuntimeProvider';
-import * as Utils from 'src/libs/utils';
+import { makeCompleteDate } from 'src/libs/utils';
 import { WorkspaceWrapper } from 'src/libs/workspace-utils';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
 import { defaultAzureWorkspace, defaultGoogleWorkspace } from 'src/testing/workspace-fixtures';
+
+import { EnvironmentNavActions, Environments, EnvironmentsProps } from './Environments';
+import { LeoResourcePermissionsProvider } from './Environments.models';
 
 jest.mock('src/libs/notifications', () => ({
   notify: jest.fn(),
@@ -151,8 +148,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(firstRuntimeRow, 3)).toBe(runtime1.labels.tool);
       expect(getTextContentForColumn(firstRuntimeRow, 5)).toBe(runtime1.status);
       expect(getTextContentForColumn(firstRuntimeRow, 6)).toBe(_.toLower(runtime1.runtimeConfig.normalizedRegion));
-      expect(getTextContentForColumn(firstRuntimeRow, 7)).toBe(Utils.makeCompleteDate(runtime1.auditInfo.createdDate));
-      expect(getTextContentForColumn(firstRuntimeRow, 8)).toBe(Utils.makeCompleteDate(runtime1.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(firstRuntimeRow, 7)).toBe(makeCompleteDate(runtime1.auditInfo.createdDate));
+      expect(getTextContentForColumn(firstRuntimeRow, 8)).toBe(makeCompleteDate(runtime1.auditInfo.dateAccessed));
     });
 
     it('Renders page correctly with multiple runtimes and workspaces', async () => {
@@ -181,8 +178,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(firstRuntimeRow, 3)).toBe(_.capitalize(runtime2.labels.tool));
       expect(getTextContentForColumn(firstRuntimeRow, 5)).toBe(runtime2.status);
       expect(getTextContentForColumn(firstRuntimeRow, 6)).toBe((runtime2.runtimeConfig as AzureConfig).region);
-      expect(getTextContentForColumn(firstRuntimeRow, 7)).toBe(Utils.makeCompleteDate(runtime2.auditInfo.createdDate));
-      expect(getTextContentForColumn(firstRuntimeRow, 8)).toBe(Utils.makeCompleteDate(runtime2.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(firstRuntimeRow, 7)).toBe(makeCompleteDate(runtime2.auditInfo.createdDate));
+      expect(getTextContentForColumn(firstRuntimeRow, 8)).toBe(makeCompleteDate(runtime2.auditInfo.dateAccessed));
 
       expect(getTextContentForColumn(tableRows[2], 0)).toBe(`${runtime1.labels.saturnWorkspaceNamespace}`);
       expect(getTextContentForColumn(tableRows[2], 1)).toBe(`${runtime1.labels.saturnWorkspaceName}`);
@@ -190,8 +187,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(tableRows[2], 3)).toBe(runtime1.labels.tool);
       expect(getTextContentForColumn(tableRows[2], 5)).toBe(runtime1.status);
       expect(getTextContentForColumn(tableRows[2], 6)).toBe(_.toLower(runtime1.runtimeConfig.normalizedRegion));
-      expect(getTextContentForColumn(tableRows[2], 7)).toBe(Utils.makeCompleteDate(runtime1.auditInfo.createdDate));
-      expect(getTextContentForColumn(tableRows[2], 8)).toBe(Utils.makeCompleteDate(runtime1.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(tableRows[2], 7)).toBe(makeCompleteDate(runtime1.auditInfo.createdDate));
+      expect(getTextContentForColumn(tableRows[2], 8)).toBe(makeCompleteDate(runtime1.auditInfo.dateAccessed));
     });
 
     it('Renders page correctly for a Dataproc Runtime', async () => {
@@ -221,11 +218,9 @@ describe('Environments', () => {
       expect(getTextContentForColumn(firstRuntimeRow, 5)).toBe(dataprocRuntime.status);
       // @ts-expect-error ignore this ts error here, we know what type of runtime config it is since it is test data (therefore, we know zone exists)
       expect(getTextContentForColumn(firstRuntimeRow, 6)).toBe(dataprocRuntime.runtimeConfig.region);
-      expect(getTextContentForColumn(firstRuntimeRow, 7)).toBe(
-        Utils.makeCompleteDate(dataprocRuntime.auditInfo.createdDate)
-      );
+      expect(getTextContentForColumn(firstRuntimeRow, 7)).toBe(makeCompleteDate(dataprocRuntime.auditInfo.createdDate));
       expect(getTextContentForColumn(firstRuntimeRow, 8)).toBe(
-        Utils.makeCompleteDate(dataprocRuntime.auditInfo.dateAccessed)
+        makeCompleteDate(dataprocRuntime.auditInfo.dateAccessed)
       );
     });
 
@@ -460,8 +455,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(firstAppRow, 3)).toBe(_.capitalize(galaxyApp.appType));
       expect(getTextContentForColumn(firstAppRow, 5)).toBe(_.capitalize(galaxyApp.status));
       expect(getTextContentForColumn(firstAppRow, 6)).toBe(galaxyApp.region);
-      expect(getTextContentForColumn(firstAppRow, 7)).toBe(Utils.makeCompleteDate(galaxyApp.auditInfo.createdDate));
-      expect(getTextContentForColumn(firstAppRow, 8)).toBe(Utils.makeCompleteDate(galaxyApp.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(firstAppRow, 7)).toBe(makeCompleteDate(galaxyApp.auditInfo.createdDate));
+      expect(getTextContentForColumn(firstAppRow, 8)).toBe(makeCompleteDate(galaxyApp.auditInfo.dateAccessed));
     });
 
     it('Renders page correctly with multiple apps and workspaces', async () => {
@@ -495,8 +490,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(firstAppRow, 3)).toBe(_.capitalize(googleApp1.appType));
       expect(getTextContentForColumn(firstAppRow, 5)).toBe(_.capitalize(googleApp1.status));
       expect(getTextContentForColumn(firstAppRow, 6)).toBe(googleApp1.region);
-      expect(getTextContentForColumn(firstAppRow, 7)).toBe(Utils.makeCompleteDate(googleApp1.auditInfo.createdDate));
-      expect(getTextContentForColumn(firstAppRow, 8)).toBe(Utils.makeCompleteDate(googleApp1.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(firstAppRow, 7)).toBe(makeCompleteDate(googleApp1.auditInfo.createdDate));
+      expect(getTextContentForColumn(firstAppRow, 8)).toBe(makeCompleteDate(googleApp1.auditInfo.dateAccessed));
 
       const secondAppRow: HTMLElement = tableRows[1];
       expect(getTextContentForColumn(secondAppRow, 0)).toBe(googleApp2.labels.saturnWorkspaceNamespace);
@@ -505,8 +500,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(secondAppRow, 3)).toBe(_.capitalize(googleApp2.appType));
       expect(getTextContentForColumn(secondAppRow, 5)).toBe(_.capitalize(googleApp2.status));
       expect(getTextContentForColumn(secondAppRow, 6)).toBe(googleApp2.region);
-      expect(getTextContentForColumn(secondAppRow, 7)).toBe(Utils.makeCompleteDate(googleApp2.auditInfo.createdDate));
-      expect(getTextContentForColumn(secondAppRow, 8)).toBe(Utils.makeCompleteDate(googleApp1.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(secondAppRow, 7)).toBe(makeCompleteDate(googleApp2.auditInfo.createdDate));
+      expect(getTextContentForColumn(secondAppRow, 8)).toBe(makeCompleteDate(googleApp1.auditInfo.dateAccessed));
 
       const thirdAppRow: HTMLElement = tableRows[2];
       expect(getTextContentForColumn(thirdAppRow, 0)).toBe(azureApp1.labels.saturnWorkspaceNamespace);
@@ -515,8 +510,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(thirdAppRow, 3)).toBe(_.capitalize(azureApp1.appType));
       expect(getTextContentForColumn(thirdAppRow, 5)).toBe(_.capitalize(azureApp1.status));
       expect(getTextContentForColumn(thirdAppRow, 6)).toBe(azureApp1.region);
-      expect(getTextContentForColumn(thirdAppRow, 7)).toBe(Utils.makeCompleteDate(azureApp1.auditInfo.createdDate));
-      expect(getTextContentForColumn(thirdAppRow, 8)).toBe(Utils.makeCompleteDate(azureApp1.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(thirdAppRow, 7)).toBe(makeCompleteDate(azureApp1.auditInfo.createdDate));
+      expect(getTextContentForColumn(thirdAppRow, 8)).toBe(makeCompleteDate(azureApp1.auditInfo.dateAccessed));
 
       const fourthAppRow: HTMLElement = tableRows[3];
       expect(getTextContentForColumn(fourthAppRow, 0)).toBe(azureApp2.labels.saturnWorkspaceNamespace);
@@ -525,8 +520,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(fourthAppRow, 3)).toBe(_.capitalize(azureApp2.appType));
       expect(getTextContentForColumn(fourthAppRow, 5)).toBe(_.capitalize(azureApp2.status));
       expect(getTextContentForColumn(fourthAppRow, 6)).toBe(azureApp2.region);
-      expect(getTextContentForColumn(fourthAppRow, 7)).toBe(Utils.makeCompleteDate(azureApp2.auditInfo.createdDate));
-      expect(getTextContentForColumn(fourthAppRow, 8)).toBe(Utils.makeCompleteDate(azureApp2.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(fourthAppRow, 7)).toBe(makeCompleteDate(azureApp2.auditInfo.createdDate));
+      expect(getTextContentForColumn(fourthAppRow, 8)).toBe(makeCompleteDate(azureApp2.auditInfo.dateAccessed));
     });
 
     it.each([
@@ -630,8 +625,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(firstDiskRow, 3)).toBe(`${disk.size}`);
       expect(getTextContentForColumn(firstDiskRow, 4)).toBe(disk.status);
       expect(getTextContentForColumn(firstDiskRow, 5)).toBe(disk.zone);
-      expect(getTextContentForColumn(firstDiskRow, 6)).toBe(Utils.makeCompleteDate(disk.auditInfo.createdDate));
-      expect(getTextContentForColumn(firstDiskRow, 7)).toBe(Utils.makeCompleteDate(disk.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(firstDiskRow, 6)).toBe(makeCompleteDate(disk.auditInfo.createdDate));
+      expect(getTextContentForColumn(firstDiskRow, 7)).toBe(makeCompleteDate(disk.auditInfo.dateAccessed));
     });
 
     it('Renders page correctly with multiple workspaces/disks', async () => {
@@ -667,8 +662,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(firstDiskRow, 3)).toBe(`${googleDisk1.size}`);
       expect(getTextContentForColumn(firstDiskRow, 4)).toBe(googleDisk1.status);
       expect(getTextContentForColumn(firstDiskRow, 5)).toBe(googleDisk1.zone);
-      expect(getTextContentForColumn(firstDiskRow, 6)).toBe(Utils.makeCompleteDate(googleDisk1.auditInfo.createdDate));
-      expect(getTextContentForColumn(firstDiskRow, 7)).toBe(Utils.makeCompleteDate(googleDisk1.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(firstDiskRow, 6)).toBe(makeCompleteDate(googleDisk1.auditInfo.createdDate));
+      expect(getTextContentForColumn(firstDiskRow, 7)).toBe(makeCompleteDate(googleDisk1.auditInfo.dateAccessed));
 
       const secondDiskRow: HTMLElement = tableRows[4];
       expect(getTextContentForColumn(secondDiskRow, 0)).toBe(googleDisk2.labels.saturnWorkspaceNamespace);
@@ -676,10 +671,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(secondDiskRow, 3)).toBe(`${googleDisk2.size}`);
       expect(getTextContentForColumn(secondDiskRow, 4)).toBe(googleDisk2.status);
       expect(getTextContentForColumn(secondDiskRow, 5)).toBe(googleDisk2.zone);
-      expect(getTextContentForColumn(secondDiskRow, 6)).toBe(Utils.makeCompleteDate(googleDisk2.auditInfo.createdDate));
-      expect(getTextContentForColumn(secondDiskRow, 7)).toBe(
-        Utils.makeCompleteDate(googleDisk2.auditInfo.dateAccessed)
-      );
+      expect(getTextContentForColumn(secondDiskRow, 6)).toBe(makeCompleteDate(googleDisk2.auditInfo.createdDate));
+      expect(getTextContentForColumn(secondDiskRow, 7)).toBe(makeCompleteDate(googleDisk2.auditInfo.dateAccessed));
 
       const thirdDiskRow: HTMLElement = tableRows[5];
       expect(getTextContentForColumn(thirdDiskRow, 0)).toBe(azureDisk1.labels.saturnWorkspaceNamespace);
@@ -687,8 +680,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(thirdDiskRow, 3)).toBe(`${azureDisk1.size}`);
       expect(getTextContentForColumn(thirdDiskRow, 4)).toBe(azureDisk1.status);
       expect(getTextContentForColumn(thirdDiskRow, 5)).toBe(azureDisk1.zone);
-      expect(getTextContentForColumn(thirdDiskRow, 6)).toBe(Utils.makeCompleteDate(azureDisk1.auditInfo.createdDate));
-      expect(getTextContentForColumn(thirdDiskRow, 7)).toBe(Utils.makeCompleteDate(azureDisk1.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(thirdDiskRow, 6)).toBe(makeCompleteDate(azureDisk1.auditInfo.createdDate));
+      expect(getTextContentForColumn(thirdDiskRow, 7)).toBe(makeCompleteDate(azureDisk1.auditInfo.dateAccessed));
 
       const fourthDiskRow: HTMLElement = tableRows[6];
       expect(getTextContentForColumn(fourthDiskRow, 0)).toBe(azureDisk2.labels.saturnWorkspaceNamespace);
@@ -696,8 +689,8 @@ describe('Environments', () => {
       expect(getTextContentForColumn(fourthDiskRow, 3)).toBe(`${azureDisk2.size}`);
       expect(getTextContentForColumn(fourthDiskRow, 4)).toBe(azureDisk2.status);
       expect(getTextContentForColumn(fourthDiskRow, 5)).toBe(azureDisk2.zone);
-      expect(getTextContentForColumn(fourthDiskRow, 6)).toBe(Utils.makeCompleteDate(azureDisk2.auditInfo.createdDate));
-      expect(getTextContentForColumn(fourthDiskRow, 7)).toBe(Utils.makeCompleteDate(azureDisk2.auditInfo.dateAccessed));
+      expect(getTextContentForColumn(fourthDiskRow, 6)).toBe(makeCompleteDate(azureDisk2.auditInfo.createdDate));
+      expect(getTextContentForColumn(fourthDiskRow, 7)).toBe(makeCompleteDate(azureDisk2.auditInfo.dateAccessed));
     });
 
     it.each([
